Extract OTP settings routes to dedicated controller module
Move the five existing OTP token management routes (settings, rename,
disable, enable, delete) from user.py to a new user_otp.py controller
module. This is a pure code extraction with no functional changes.

The new module is auto-discovered via import_all("noggin.controller")
in app.py. Also fix the localized Markup string in _user_mod and
remove a redundant username reassignment in user_settings_otp_delete.
